requirements.txt

fastapi>=0.115.0
uvicorn>=0.30.0
openai>=1.40.0
pydantic>=2.8.0
python-dotenv>=1.0.0
